requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.7.0
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
feedparser==6.0.11
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
load-dotenv==0.1.0
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
